+++ app/Commands/Download.php
@@ -6,7 +6,6 @@ use Carbon\Carbon;
 use Carbon\CarbonInterval;
 use Illuminate\Console\Scheduling\Schedule;
 use Illuminate\Contracts\Process\ProcessResult;
-use Illuminate\Support\Facades\File;
 use Illuminate\Support\Facades\Process;
 use Illuminate\Support\Facades\Storage;
 use Illuminate\Support\Number;
@@ -111,30 +110,8 @@ class Download extends BaseCommand
             }
         }
 
-        $includeServers = null;
-        $excludeServers = null;
-
-        $include = $this->option('include');
-        if (!empty($include))
-        {
-            if (!File::exists($include))
-            {
-                $this->fail("Include file [{$include}] does not exists or is not readable");
-            }
-
-            $includeServers = array_filter(explode(PHP_EOL, File::get($include)));
-        }
-
-        $exclude = $this->option('exclude');
-        if (!empty($exclude))
-        {
-            if (!File::exists($exclude))
-            {
-                $this->fail("Exclude file [{$exclude}] does not exists or is not readable");
-            }
-
-            $excludeServers = array_filter(explode(PHP_EOL, File::get($exclude)));
-        }
+        $includeServers = $this->serverList('include');
+        $excludeServers = $this->serverList('exclude');
 
         $failed = collect($servers)
             ->filter(function ($server) use ($includeServers) {